Improve memory usage when loading models with multiple pipeline stages
We make a custom PipelineModule class, which splits the initialization
into different methods. Using this, we can load and quantize the model
one process at a time, move it to CPU RAM, then free layers that won't
be needed by that process. This lets me load llama2 70b in 4 stages,
with 128GB system RAM.