Refactor dataset_utils.py
Use the Huggingface Datasets library for all transformations on the
dataset. This simplifies the code and is more idiomatic. It also means
we no longer need to manually handle caching: Datasets does that for us,
caching the results of each transformation to disk and loading from
cache if nothing has changed.