merge_lora: switch copy order and add tqdm
The order switch is done as there are often big quant files in the input directory that are inadvertently copied over to the new location (gguf files in particular). By copying first, the user is able to abort and redo the process with the files cleaned out, without having to wait for the merge to finish before finding out.