Allow manually saving checkpoint (and optionally quitting)
This adds a simple mechanism to manually checkpoint, and optionally
quit, by writing either a 'save' or 'save_quit' file to the run_dir.
Long term, it might be better to intercept Ctrl+C and have some kind of
"command prompt". But that is harder to implement, and I'm not even sure
it could work with all of the multiple processes that are happening
(deepspeed launcher + 1 process for each GPU). This is way is simple and
straightforward. On linux it's as easy as a "touch save_quit" in the
run_dir.