do not crash if eval_size is not set
For configurations where eval dataset is explicitly set, the eval_size is not used or set to 0. In both of these cases, due to the if case regarding eval_size == 0 as not True, the code will fail to set the val_set_size cfg param, resulting in a crash in Axolotl code, which assumes it is a numeric non-None value.