Pipeline parallel sampling schedule
Still a work in progress, but a minimal example works. Will be used for
future training techniques that require sampling from the model.

This change kind of spiraled out of control and led to a lot of refactoring.
I tested all model types except Cohere, Phi3, Gemma2. If something broke,
it should be a relatively easy fix.