Revert attention_mask handling in input layer
What I was trying to do is not sound in general for pipeline parallelism.
Depending on how the model processes the attention mask, it could be a
tensor in one micro batch, but then None in the next. If we set None to
an empty tensor to signal that we should convert it back to None later,
we no longer have static tensor sizes and things will break in strange
ways.