Always use dynamic_shape=True in the pipeline engine, and support configuring attn_implementation
Motivated by the rejected sampling feature. For sampling only, SDPA is significantly faster than
flash attention because the attention mask tensor is explicitly passed. So we want to be able to
set SDPA. But then, SDPA only works if we can pass an empty tensor between layers to signal the None
attention_mask. And that requires dynamic_shape=True.

When training llama3 8b at 4096 sequence length, I'm not able to measure any performance
difference between dynamic_shape True and False. So we'll simplify the code by just always
having it as True.